requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
bcrypt
pyjwt
python-multipart
sqlalchemy
jinja2